requirements.txt
SpeechRecognition==3.10.0
pyttsx3==2.90
wikipedia-api==0.6.0
pyaudio==0.2.14